+++ packages/utils/test/deepVisit.spec.ts
@@ -1,4 +1,5 @@
 import deepVisit from '../src/deepVisit';
+import DeepVisitBreak from '../src/DeepVisitBreak';
 
 it('should visit all properties.', () => {
   const obj = {
@@ -18,6 +19,7 @@ it('should visit all properties.', () => {
   const result: any[] = [];
   deepVisit(obj, (_, k, po, keyPathList) => {
     result.push(keyPathList.join('.'));
+    return DeepVisitBreak.NO;
   });
   expect(result).toEqual([
     'name',
@@ -37,6 +39,39 @@ it('should handle the circle.', () => {
   const result: any[] = [];
   deepVisit(obj, (_, k, po, keyPathList) => {
     result.push(keyPathList.join('.'));
+    return DeepVisitBreak.NO;
   });
   expect(result).toEqual(['name', 'c']);
+});
+
+it('should break all from visiting.', () => {
+  const obj = {
+    name: 'yujiang',
+    age: 20,
+  };
+  const result: any[] = [];
+  deepVisit(obj, (_, k, po, keyPathList) => {
+    result.push(keyPathList.join('.'));
+    return DeepVisitBreak.ALL;
+  });
+  expect(result).toEqual(['name']);
+});
+
+it('should break children from visiting.', () => {
+  const obj = {
+    name: 'yujiang',
+    age: 20,
+    address: {
+      city: 'chengdu',
+    },
+  };
+  const result: any[] = [];
+  deepVisit(obj, (_, k, po, keyPathList) => {
+    result.push(keyPathList.join('.'));
+    if (k === 'address') {
+      return DeepVisitBreak.CHILDREN;
+    }
+    return DeepVisitBreak.NO;
+  });
+  expect(result).toEqual(['name', 'age', 'address']);
 });