+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 - [+] **@goldfishjs/utils:**
   - visit the object deeply and handle the circular reference.
 
+### 🐞 Bug Fixes
+
+- [^] **@goldfishjs/reactive:**
+  - use `deepVisit()` to handle circular objects.
+
 ## 1.1.3
 
 ### 🐞 Bug Fixes